test(#1146): address Copilot review — move seeding helper out of conftest
- Move SeededParsedStatement + seed_parsed_statement into tests/factories.py
  (a tests.* importable package) so the test imports the type from there
  instead of from conftest (tests/e2e has no __init__.py).
- Use base_date + timedelta(days=index) for txn_date so seeding >21
  transactions no longer overflows the month (reusable-helper robustness).
- Fix the test module docstring grammar.
- Update the e2e conftest module docstring to reflect that it now also hosts
  the API-tier seeded fixture, not only Playwright fixtures.
