+++ apps/backend/tests/factories.py
@@ -14,7 +14,8 @@ Usage:
     account = await AccountFactory.create_async(db, name="Cash")
 """
 
-from datetime import UTC, date, datetime
+from dataclasses import dataclass, field
+from datetime import UTC, date, datetime, timedelta
 from decimal import Decimal
 from typing import TypeVar
 from uuid import UUID, uuid4
@@ -27,7 +28,7 @@ from src.models.journal import Direction, JournalEntry, JournalEntryStatus, Jour
 from src.models.layer1 import DocumentType, UploadedDocument
 from src.models.layer2 import AtomicTransaction, TransactionDirection
 from src.models.reconciliation import ReconciliationMatch, ReconciliationStatus
-from src.models.statement_enums import BankStatementStatus
+from src.models.statement_enums import BankStatementStatus, Stage1Status
 from src.models.statement_summary import StatementSummary
 from src.models.user import User
 
@@ -249,3 +250,137 @@ class ReconciliationMatchFactory(factory.Factory, AsyncFactoryMixin):
     @classmethod
     def _build_kwargs(cls, atomic_txn_id: UUID, journal_entry_ids: list[str], **kwargs) -> dict:
         return {"atomic_txn_id": atomic_txn_id, "journal_entry_ids": journal_entry_ids, **kwargs}
+
+
+# ---------------------------------------------------------------------------
+# Seeded no-LLM parsed statement (EPIC-008 / AC8.21)
+# ---------------------------------------------------------------------------
+
+
+@dataclass
+class SeededParsedStatement:
+    """Handle for a fixture-seeded, already-parsed statement.
+
+    Carries the layered DWD records (ODS document, conform envelope, atomic
+    transactions) injected directly into the test database so the downstream
+    review -> reconcile -> report journeys can run with **zero provider calls**.
+    The provider/LLM extraction seam (``ExtractionService.parse_document`` ->
+    ``stream_ai_json``) is bypassed entirely: the parsed result is materialized
+    by hand, exactly as the extraction pipeline would have persisted it.
+    """
+
+    user_id: UUID
+    document: UploadedDocument
+    statement: StatementSummary
+    transactions: list[AtomicTransaction] = field(default_factory=list)
+
+    @property
+    def id(self) -> UUID:
+        """The statement (DWD conform) id, as the statements API addresses it."""
+        return self.statement.id
+
+    @property
+    def original_filename(self) -> str:
+        """The ODS filename the statement-list row link renders (#1142)."""
+        return self.document.original_filename
+
+
+async def seed_parsed_statement(
+    db: AsyncSession,
+    user_id: UUID,
+    *,
+    institution: str = "DBS",
+    original_filename: str = "dbs_statement_2026_01.pdf",
+    opening_balance: Decimal = Decimal("1000.00"),
+    transactions: list[dict] | None = None,
+) -> SeededParsedStatement:
+    """Inject an already-parsed statement into the database (no provider call).
+
+    Seeds the three layered records the extraction pipeline would have written:
+
+    * ODS ``UploadedDocument`` — carries ``original_filename`` (the field the
+      statement-list stretched-link row renders; empty during real parsing,
+      which is the #1142 invisible-link bug).
+    * DWD ``StatementSummary`` — ``status=PARSED`` / ``stage1_status=PENDING_REVIEW``
+      conform envelope, linked to the ODS document.
+    * Layer-2 ``AtomicTransaction`` rows — joined back to the statement via
+      ``source_documents[*].doc_id == uploaded_document_id`` (the exact join
+      ``resolve_statement_transactions`` performs).
+
+    Returns a :class:`SeededParsedStatement` handle. All amounts use ``Decimal``
+    (never ``float``) per the monetary red line.
+    """
+    rows = (
+        transactions
+        if transactions is not None
+        else [
+            {"description": "Salary", "amount": Decimal("5000.00"), "direction": TransactionDirection.IN},
+            {"description": "Coffee Shop", "amount": Decimal("5.00"), "direction": TransactionDirection.OUT},
+        ]
+    )
+
+    file_hash = uuid4().hex + uuid4().hex[:32]
+    document = await UploadedDocumentFactory.create_async(
+        db,
+        user_id=user_id,
+        file_hash=file_hash,
+        original_filename=original_filename,
+        document_type=DocumentType.BANK_STATEMENT,
+    )
+
+    # Compute the closing balance from the seeded movements so the balance chain
+    # (open + ΣIN − ΣOUT ≈ close) ties out without any provider involvement.
+    closing_balance = opening_balance
+    for row in rows:
+        if row["direction"] == TransactionDirection.IN:
+            closing_balance += row["amount"]
+        else:
+            closing_balance -= row["amount"]
+
+    statement = await StatementSummaryFactory.create_async(
+        db,
+        user_id=user_id,
+        file_hash=file_hash,
+        uploaded_document_id=document.id,
+        institution=institution,
+        currency="SGD",
+        period_start=date(2026, 1, 1),
+        period_end=date(2026, 1, 31),
+        opening_balance=opening_balance,
+        closing_balance=closing_balance,
+        status=BankStatementStatus.PARSED,
+        stage1_status=Stage1Status.PENDING_REVIEW,
+        confidence_score=95,
+    )
+
+    doc_marker = {"doc_id": str(document.id), "doc_type": DocumentType.BANK_STATEMENT.value}
+    base_date = date(2026, 1, 10)
+    seeded_txns: list[AtomicTransaction] = []
+    running_balance = opening_balance
+    for index, row in enumerate(rows):
+        if row["direction"] == TransactionDirection.IN:
+            running_balance += row["amount"]
+        else:
+            running_balance -= row["amount"]
+        txn = await AtomicTransactionFactory.create_async(
+            db,
+            user_id=user_id,
+            source_documents=[doc_marker],
+            # timedelta keeps the date valid for any transaction count (avoids
+            # day-overflow once index pushes past the month end).
+            txn_date=base_date + timedelta(days=index),
+            description=row["description"],
+            amount=row["amount"],
+            direction=row["direction"],
+            currency="SGD",
+            balance_after=running_balance,
+        )
+        seeded_txns.append(txn)
+
+    await db.commit()
+    return SeededParsedStatement(
+        user_id=user_id,
+        document=document,
+        statement=statement,
+        transactions=seeded_txns,
+    )
